refactor(record-entry): dedupe page validation and fix status icon click
Share getPageValidationStatus via recordPageValidation, and treat SVG
status-icon clicks as label clicks so expanded rows do not collapse.
